blogs routes added

+++ server/adoption/src/controller/adoption.controller.ts
@@ -42,11 +42,16 @@ export const getPostWithId = async (req: Request, res: Response): Promise<any> =
     }
 }
 
-export const deletePost = async (req: Request, res: Response): Promise<any> => {
+export const deletePost = async (req: any, res: Response): Promise<any> => {
     try {
+
         const { id } = req.params
         if (!id || mongoose.Types.ObjectId.isValid(JSON.stringify(id))) {
             return res.status(400).json({ error: 'invalid id' })
+        }
+        const post = await Post.findById(id)
+        if(post?.userId!=req.user.userId){
+            return res.status(401).json({msg:"unauthorized"})
         }
         await Post.findByIdAndDelete(id)
         return res.status(200).json({msg:`post with id: ${id} has been deleted`})
